+++ R/s3.generics.R
@@ -193,6 +193,7 @@ hypothesis <- function(x, hypothesis, class = "b", alpha = 0.05, ...)
 #' @param x An \code{R} object typically of class \code{brmsfit}
 #' @param parameters Name of parameters for which posterior samples should be returned, as given by a character vector or regular expressions.
 #'   By default, all posterior samples of all parameters are extracted
+#' @param add.chains A flag indicating if the returned data.frame should contain information on the chains
 #' @param ... Currently ignored
 #'   
 #' @details Currently there are methods for \code{brmsfit} objects.
@@ -215,7 +216,7 @@ hypothesis <- function(x, hypothesis, class = "b", alpha = 0.05, ...)
 #' }
 #' 
 #' @export 
-posterior.samples <- function(x, parameters = NA, ...)
+posterior.samples <- function(x, parameters = NA, add.chains = FALSE,...)
   UseMethod("posterior.samples")
 
 #' Extract parameter names